requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
streamlit
langchain-experimental
sentence-transformers
langchain-chroma
langchainhub
pypdf
rapidocr-onnxruntime